model_partitioner/requirements.txt
torch>=2.0.0
transformers>=4.35.0
safetensors>=0.4.0
onnx>=1.15.0
onnxruntime>=1.16.0
onnxruntime-gpu>=1.16.0  # Optional: for GPU support
Pillow>=10.0.0
numpy>=1.24.0
psutil>=5.9.0
accelerate>=0.24.0
bitsandbytes>=0.41.0  # For quantization

